@@ -8,19 +8,20 @@
         "erusev/parsedown-extra": "^0.8.1",
         "jajo/jsondb": "^3.0",
         "ocramius/package-versions": "^2.5",
-        "rector/rector": "dev-main",
-        "symfony/asset": "^6.1",
-        "symfony/cache": "^6.1",
-        "symfony/dependency-injection": "^6.1",
-        "symfony/dotenv": "^6.1",
-        "symfony/form": "^6.1",
-        "symfony/framework-bundle": "^6.1",
-        "symfony/http-kernel": "^6.1",
-        "symfony/process": "^6.1",
-        "symfony/twig-bundle": "^6.1",
-        "symfony/uid": "^6.1",
-        "symfony/webpack-encore-bundle": "^1.13",
-        "symfony/yaml": "^6.1",
+        "rector/rector": "^0.15.11",
+        "sebastian/diff": "5.0 as 4.0.4",
+        "symfony/asset": "^6.2",
+        "symfony/cache": "^6.2",
+        "symfony/dependency-injection": "^6.2",
+        "symfony/dotenv": "^6.2",
+        "symfony/form": "^6.2",
+        "symfony/framework-bundle": "^6.2",
+        "symfony/http-kernel": "^6.2",
+        "symfony/process": "^6.2",
+        "symfony/twig-bundle": "^6.2",
+        "symfony/uid": "^6.2",
+        "symfony/webpack-encore-bundle": "^1.16",
+        "symfony/yaml": "^6.2",
         "symplify/autowire-array-parameter": "^11.1",
         "symplify/package-builder": "^11.1",
         "symplify/smart-file-system": "^11.1",
@@ -32,7 +33,8 @@
         "phpstan/phpstan": "^1.8",
         "phpunit/phpunit": "^10.0",
         "roave/security-advisories": "dev-master",
-        "symfony/debug-bundle": "^6.1",
+        "symfony/browser-kit": "^6.2",
+        "symfony/debug-bundle": "^6.2",
         "symplify/coding-standard": "^11.1",
         "symplify/easy-ci": "^11.1",
         "symplify/easy-coding-standard": "^11.1",
@@ -41,7 +43,8 @@
         "symplify/phpstan-rules": "^11.1",
         "tomasvotruba/cognitive-complexity": "^0.0.5",
         "tomasvotruba/type-coverage": "^0.0.7",
-        "tomasvotruba/unused-public": "^0.0.33"
+        "tomasvotruba/unused-public": "^0.0.33",
+        "tracy/tracy": "^2.9"
     },
     "autoload": {
         "psr-4": {
